useReactionsResource hook
